requirements.txt
pyTelegramBotAPI==4.26.0
requests==2.31.0
yandexgpt-python==0.2.3
